Teams Page now showing Players
/Teams page now correctly displays the team's player (if any) and shows their picture (if no picture -> default)

+++ Balkana/Data/Models/PlayerStatistic_CS2.cs
@@ -13,9 +13,9 @@ namespace Balkana.Data.Models
         public int Deaths { get; set; }
         public int KAST { get; set; }
         public int HSkills { get; set; }
-        public int HLTV3 { get; set; }
-        public int HLTV2 { get; set; }
-        public int HLTV1 { get; set; }
+        public double HLTV3 { get; set; }
+        public double HLTV2 { get; set; }
+        public double HLTV1 { get; set; }
         public int UD { get; set; }
         public int FK { get; set; }
         public int FD { get; set; }
